using Double in string2number.hs

+++ haskell/day-2/string2number.hs
@@ -1,5 +1,5 @@
 module StringToNum where
 
-  convertToNum :: String -> Float
+  convertToNum :: String -> Double
   convertToNum s = read cs
     where cs = filter (`elem` '.':['0'..'9']) s
